Move generator unit tests into the test/unit/generators dir
