+++ packages/nx-plugin/test/unit/generators/circleciGenerator.spec.ts
@@ -1,7 +1,7 @@
 import { Tree } from "@nrwl/devkit";
 import { createTreeWithEmptyWorkspace } from "@nrwl/devkit/testing";
 
-import { circleciGenerator } from "../../src/generators";
+import { circleciGenerator } from "../../../src/generators/index";
 
 describe("circleciGenerator", () => {
   let tree: Tree;
