+++ packages/nx-plugin/test/unit/generators/eslintGenerator.spec.ts
@@ -3,7 +3,7 @@ import { createTreeWithEmptyWorkspace } from "@nrwl/devkit/testing";
 import { Linter } from "eslint";
 import { PackageJson } from "nx/src/utils/package-json";
 
-import { eslintGenerator } from "../../src/generators/linting";
+import { eslintGenerator } from "../../../src/generators/linting";
 
 describe("eslintGenerator", () => {
   let tree: Tree;
