+++ packages/nx-plugin/test/unit/generators/lintStagedGenerator.spec.ts
@@ -2,7 +2,7 @@ import { readJson, Tree } from "@nrwl/devkit";
 import { createTreeWithEmptyWorkspace } from "@nrwl/devkit/testing";
 import { PackageJson } from "nx/src/utils/package-json";
 
-import { lintStagedGenerator } from "../../src/generators/linting/lintStaged";
+import { lintStagedGenerator } from "../../../src/generators/linting/lintStaged";
 
 describe("lintStagedGenerator", () => {
   let tree: Tree;
